+++ R/autoplot.R
@@ -300,8 +300,7 @@ plot_metrics <- function(
   ) +
     ggplot2::aes(
       x = .data[["CartridgeID"]],
-      y = .data[[x]],
-      colour = .data[[colour]]
+      y = .data[[x]]
     ) +
     ggplot2::scale_colour_viridis_d(option = "plasma", direction = 1, end = 0.85) +
     ggplot2::geom_boxplot(
@@ -316,6 +315,7 @@ plot_metrics <- function(
         list(
           ggplot2::geom_point(
             data = ~ .x[!(is_outlier)],
+            mapping = ggplot2::aes(colour = .data[[colour]]),
             size = size, na.rm = TRUE,
             position = ggplot2::position_jitter(width = 0.25, height = 0)
           ),
@@ -337,6 +337,7 @@ plot_metrics <- function(
         )
       } else {
         ggplot2::geom_point(
+          mapping = ggplot2::aes(colour = .data[[colour]]),
           size = size, na.rm = TRUE,
           position = ggplot2::position_jitter(width = 0.25, height = 0)
         )
@@ -432,8 +433,7 @@ plot_cg <- function(
   ) +
     ggplot2::aes(
       x = .data[["Name"]],
-      y = .data[["Count"]] + 1,
-      colour = .data[[colour]]
+      y = .data[["Count"]] + 1
     ) +
     ggplot2::scale_colour_viridis_d(option = "plasma", direction = 1, end = 0.85) +
     ggplot2::geom_boxplot(
@@ -448,6 +448,7 @@ plot_cg <- function(
         list(
           ggplot2::geom_point(
             data = ~ .x[!(is_outlier)],
+            mapping = ggplot2::aes(colour = .data[[colour]]),
             size = size, na.rm = TRUE,
             position = ggplot2::position_jitter(width = 0.25, height = 0)
           ),
@@ -469,13 +470,13 @@ plot_cg <- function(
         )
       } else {
         ggplot2::geom_point(
+          mapping = ggplot2::aes(colour = .data[[colour]]),
           size = size, na.rm = TRUE,
           position = ggplot2::position_jitter(width = 0.25, height = 0)
         )
       }
     } +
     ggplot2::scale_y_log10(
-      # limits = c(1, NA),
       labels = function(x) format(x, big.mark = ",")
     ) +
     ggplot2::labs(
@@ -532,7 +533,6 @@ plot_pn <- function(
     ggplot2::geom_line() +
     ggplot2::facet_wrap(facets = "CodeClass", scales = "free_y", ncol = 2) +
     ggplot2::scale_y_log10(
-      # limits = c(1, NA),
       labels = function(x) format(x, big.mark = ",")
     ) +
     ggplot2::scale_x_discrete(labels = NULL) +
@@ -919,11 +919,11 @@ plot_pfnf <- function(
       }
     } +
     ggplot2::labs(x = "Negative Factor", y = "Positive Factor", colour = colour) +
-    ggplot2::scale_y_log10() +
+    ggplot2::scale_y_continuous(transform = transform_log10_infinite()) +
     ggplot2::geom_rect(
       data = data.table::data.table(
         ymin = nacho_object$outliers_thresholds[["Positive_factor"]],
-        ymax = c(0, Inf)
+        ymax = c(-Inf, Inf)
       ),
       mapping = ggplot2::aes(xmin = -Inf, xmax = Inf, ymin = .data[["ymin"]], ymax = .data[["ymax"]]),
       fill = "#b22222",
@@ -1024,14 +1024,14 @@ plot_hf <- function(
       }
     } +
     ggplot2::labs(x = "Positive Factor", y = "Housekeeping Factor", colour = colour) +
-    ggplot2::scale_x_log10() +
-    ggplot2::scale_y_log10() +
+    ggplot2::scale_x_continuous(transform = transform_log10_infinite()) +
+    ggplot2::scale_y_continuous(transform = transform_log10_infinite()) +
     ggplot2::geom_rect(
       data = data.table::data.table(
-        xmin = c(0, 0, nacho_object$outliers_thresholds[["Positive_factor"]]),
-        xmax = c(Inf, Inf, 0, Inf),
-        ymin = c(nacho_object$outliers_thresholds[["House_factor"]], 0, 0),
-        ymax = c(0, Inf, Inf, Inf)
+        xmin = c(-Inf, -Inf, nacho_object$outliers_thresholds[["Positive_factor"]]),
+        xmax = c(Inf, Inf, -Inf, Inf),
+        ymin = c(nacho_object$outliers_thresholds[["House_factor"]], -Inf, -Inf),
+        ymax = c(-Inf, Inf, Inf, Inf)
       ),
       mapping = ggplot2::aes(
         xmin = .data[["xmin"]],
@@ -1136,7 +1136,6 @@ plot_norm <- function(
     ggplot2::scale_colour_viridis_d(option = "plasma", direction = 1, end = 0.85) +
     ggplot2::scale_x_discrete(label = NULL) +
     ggplot2::scale_y_log10(
-      # limits = c(1, NA),
       labels = function(x) format(x, big.mark = ",")
     ) +
     ggplot2::labs(