+++ R/deploy.R
@@ -3,6 +3,8 @@
 #' @param directory [[character]] A character vector of one path to the new location.
 #' @param app_name [[character]] A character vector defining the shiny application name in the new location.
 #'
+#' @details The app needs the markdown package on the server to show its help pages.
+#'
 #' @return [[logical]] A logical indicating whether the deployment is successfull (`TRUE`) or not (`FALSE`).
 #' @export
 #'